// frontend/src/components/placement/PlacementProgressTracker.d.ts
import React from 'react';
import { PlacementRoundSession } from '../../types/placement';
interface PlacementProgressTrackerProps {
    rounds: PlacementRoundSession[];
    currentRoundOrder: number;
}
export declare const PlacementProgressTracker: React.FC<PlacementProgressTrackerProps>;
export {};
//# sourceMappingURL=PlacementProgressTracker.d.ts.map